Factorized ResNet and DenseNet models
